feat: Refactor Users api view This PR moves the ORM logic from MySQL api to model methods in favor of a cleaner code.
