+++ forum/backends/mysql/api.py
@@ -2,19 +2,16 @@
 
 import math
 import random
-from datetime import timedelta
 from typing import Any, Optional, Union
 
 from django.contrib.auth.models import User  # pylint: disable=E5142
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ObjectDoesNotExist
-from django.core.paginator import Paginator
 from django.db.models import (
     Count,
     Exists,
     F,
     IntegerField,
-    Max,
     OuterRef,
     Q,
     Subquery,
@@ -40,7 +37,6 @@ from forum.backends.mysql.models import (
     UserVote,
     Commentable,
 )
-from forum.constants import RETIRED_BODY, RETIRED_TITLE
 from forum.utils import get_group_ids_from_params
 
 
@@ -815,22 +811,18 @@ class MySQLBackend(AbstractBackend):
         return presenters
 
     @staticmethod
-    def get_username_from_id(user_id: str) -> Optional[str]:
-        """
-        Retrieve the username associated with a given user ID.
+    def get_user(user_id: str) -> dict[str, Any] | None:
+        """Return user from user_id."""
+        forum_user = ForumUser.get_by_user_id(user_id)
+        return forum_user.to_dict() if forum_user else None
 
-        Args:
-            _id (int): The unique identifier of the user.
-
-        Returns:
-            Optional[str]: The username of the user if found, or None if not.
-
-        """
-        try:
-            user = User.objects.get(pk=user_id)
-        except ObjectDoesNotExist:
+    @staticmethod
+    def get_user_by_username(username: str | None) -> dict[str, Any] | None:
+        """Return user from username."""
+        if not username:
             return None
-        return user.username
+        forum_user = ForumUser.get_by_username(username)
+        return forum_user.to_dict() if forum_user else None
 
     @staticmethod
     def validate_object(model: str, obj_id: str) -> Any:
@@ -1020,84 +1012,14 @@ class MySQLBackend(AbstractBackend):
     def user_to_hash(
         cls, user_id: str, params: Optional[dict[str, Any]] = None
     ) -> dict[str, Any]:
-        """
-        Converts user data to a hash
-        """
-        user = User.objects.get(pk=user_id)
-        forum_user = ForumUser.objects.get(user__pk=user_id)
-        if params is None:
-            params = {}
-
-        user_data = forum_user.to_dict()
-        hash_data = {}
-        hash_data["username"] = user_data["username"]
-        hash_data["external_id"] = user_data["external_id"]
-        hash_data["id"] = user_data["external_id"]
-
-        if params.get("complete"):
-            subscribed_thread_ids = cls.find_subscribed_threads(user_id)
-            upvoted_ids = cls.get_user_voted_ids(user_id, "up")
-            downvoted_ids = cls.get_user_voted_ids(user_id, "down")
-            hash_data.update(
-                {
-                    "subscribed_thread_ids": subscribed_thread_ids,
-                    "subscribed_commentable_ids": [],
-                    "subscribed_user_ids": [],
-                    "follower_ids": [],
-                    "id": user_id,
-                    "upvoted_ids": upvoted_ids,
-                    "downvoted_ids": downvoted_ids,
-                    "default_sort_key": user_data["default_sort_key"],
-                }
-            )
-
-        if params.get("course_id"):
-            threads = CommentThread.objects.filter(
-                author=user,
-                course_id=params["course_id"],
-                anonymous=False,
-                anonymous_to_peers=False,
-            )
-            comments = Comment.objects.filter(
-                author=user,
-                course_id=params["course_id"],
-                anonymous=False,
-                anonymous_to_peers=False,
-            )
-            comment_ids = list(comments.values_list("pk", flat=True))
-            if params.get("group_ids"):
-                group_threads = threads.filter(
-                    group_id__in=params["group_ids"] + [None]
-                )
-                group_thread_ids = [str(thread.pk) for thread in group_threads]
-                threads_count = len(group_thread_ids)
-                comment_thread_ids = cls.filter_standalone_threads(comment_ids)
-
-                group_comment_threads = CommentThread.objects.filter(
-                    id__in=comment_thread_ids, group_id__in=params["group_ids"] + [None]
-                )
-                group_comment_thread_ids = [
-                    str(thread.pk) for thread in group_comment_threads
-                ]
-                comments_count = sum(
-                    1
-                    for comment_thread_id in comment_thread_ids
-                    if comment_thread_id in group_comment_thread_ids
-                )
-            else:
-                thread_ids = [str(thread.pk) for thread in threads]
-                threads_count = len(thread_ids)
-                comment_thread_ids = cls.filter_standalone_threads(comment_ids)
-                comments_count = len(comment_thread_ids)
-
-            hash_data.update(
-                {
-                    "threads_count": threads_count,
-                    "comments_count": comments_count,
-                }
-            )
-
-        return hash_data
+        """Converts user data to a hash."""
+        return ForumUser.user_to_hash(
+            user_id,
+            params,
+            filter_standalone_threads=cls.filter_standalone_threads,
+            find_subscribed_threads=cls.find_subscribed_threads,
+            get_user_voted_ids=cls.get_user_voted_ids,
+        )
 
     @staticmethod
     def replace_username(user_id: str, username: str) -> None:
@@ -1112,24 +1034,12 @@ class MySQLBackend(AbstractBackend):
     @staticmethod
     def unsubscribe_all(user_id: str) -> None:
         """Unsubscribe user from all content."""
-        Subscription.unsubscribe_all(user_id)
+        ForumUser.unsubscribe_all(user_id)
 
-    # Kept method signature same as mongo implementation
     @staticmethod
-    def retire_all_content(
-        user_id: str, username: str
-    ) -> None:  # pylint: disable=W0613
+    def retire_all_content(user_id: str) -> None:
         """Retire all content from user."""
-        comments = Comment.objects.filter(author__pk=user_id)
-        for comment in comments:
-            comment.body = RETIRED_BODY
-            comment.save()
-
-        comment_threads = CommentThread.objects.filter(author__pk=user_id)
-        for comment_thread in comment_threads:
-            comment_thread.body = RETIRED_BODY
-            comment_thread.title = RETIRED_TITLE
-            comment_thread.save()
+        ForumUser.retire_all_content(user_id)
 
     @staticmethod
     def find_or_create_read_state(user_id: str, thread_id: str) -> dict[str, Any]:
@@ -1199,120 +1109,12 @@ class MySQLBackend(AbstractBackend):
     @classmethod
     def build_course_stats(cls, author_id: str, course_id: str) -> None:
         """Build course stats."""
-        author = User.objects.get(pk=author_id)
-        threads = CommentThread.objects.filter(
-            author=author,
-            course_id=course_id,
-            anonymous_to_peers=False,
-            anonymous=False,
-        )
-        comments = Comment.objects.filter(
-            author=author,
-            course_id=course_id,
-            anonymous_to_peers=False,
-            anonymous=False,
-        )
-
-        responses = comments.filter(parent__isnull=True)
-        replies = comments.filter(parent__isnull=False)
-
-        comment_ids = [comment.pk for comment in comments]
-        threads_ids = [thread.pk for thread in threads]
-
-        active_flags_comments = (
-            AbuseFlagger.objects.filter(
-                content_object_id__in=comment_ids, content_type=Comment().content_type
-            )
-            .values("content_object_id")
-            .annotate(count=Count("content_object_id"))
-            .count()
-        )
-
-        active_flags_threads = (
-            AbuseFlagger.objects.filter(
-                content_object_id__in=threads_ids,
-                content_type=CommentThread().content_type,
-            )
-            .values("content_object_id")
-            .annotate(count=Count("content_object_id"))
-            .count()
-        )
-
-        active_flags = active_flags_comments + active_flags_threads
-
-        inactive_flags_comments = (
-            HistoricalAbuseFlagger.objects.filter(
-                content_object_id__in=comment_ids, content_type=Comment().content_type
-            )
-            .values("content_object_id")
-            .annotate(count=Count("content_object_id"))
-            .count()
-        )
-
-        inactive_flags_threads = (
-            HistoricalAbuseFlagger.objects.filter(
-                content_object_id__in=threads_ids,
-                content_type=CommentThread().content_type,
-            )
-            .values("content_object_id")
-            .annotate(count=Count("content_object_id"))
-            .count()
-        )
-
-        inactive_flags = inactive_flags_comments + inactive_flags_threads
-
-        threads_updated_at = threads.aggregate(Max("updated_at"))["updated_at__max"]
-        comments_updated_at = comments.aggregate(Max("updated_at"))["updated_at__max"]
-
-        updated_at = max(
-            threads_updated_at or timezone.now() - timedelta(days=365 * 100),
-            comments_updated_at or timezone.now() - timedelta(days=365 * 100),
-        )
-
-        stats, _ = CourseStat.objects.get_or_create(user=author, course_id=course_id)
-        stats.threads = threads.count()
-        stats.responses = responses.count()
-        stats.replies = replies.count()
-        stats.active_flags = active_flags
-        stats.inactive_flags = inactive_flags
-        stats.last_activity_at = updated_at
-        stats.save()
-        cls.update_user_stats_for_course(author_id, stats.to_dict())
+        ForumUser.build_course_stats(author_id, course_id)
 
     @classmethod
     def update_all_users_in_course(cls, course_id: str) -> list[str]:
         """Update all user stats in a course."""
-        course_comments = Comment.objects.filter(
-            anonymous=False,
-            anonymous_to_peers=False,
-            course_id=course_id,
-        )
-        course_threads = CommentThread.objects.filter(
-            anonymous=False,
-            anonymous_to_peers=False,
-            course_id=course_id,
-        )
-
-        comment_authors = set(course_comments.values_list("author__id", flat=True))
-        thread_authors = set(course_threads.values_list("author__id", flat=True))
-        author_ids = list(comment_authors | thread_authors)
-
-        for author_id in author_ids:
-            cls.build_course_stats(author_id, course_id)
-        return author_ids
-
-    @staticmethod
-    def get_user_by_username(username: str | None) -> dict[str, Any] | None:
-        """Return user from username."""
-        try:
-            user = User.objects.get(username=username)
-        except User.DoesNotExist:
-            return None
-        try:
-            forum_user = ForumUser.objects.get(user=user)
-        except ForumUser.DoesNotExist:
-            return None
-        return forum_user.to_dict()
+        return ForumUser.update_all_users_in_course(course_id)
 
     @staticmethod
     def generate_id() -> str:
@@ -1326,20 +1128,7 @@ class MySQLBackend(AbstractBackend):
         default_sort_key: Optional[str] = "date",
     ) -> str:
         """Find or create user."""
-        username = username or user_id
-        try:
-            user = User.objects.get(pk=int(user_id))
-        except User.DoesNotExist:
-            user = None
-
-        if user is None:
-            if User.objects.filter(username=username).exists():
-                raise ValueError(f"User with username {username} already exists")
-            user = User.objects.create(pk=int(user_id), username=username)
-
-        forum_user, _ = ForumUser.objects.get_or_create(
-            user=user, defaults={"default_sort_key": default_sort_key}
-        )
+        forum_user = ForumUser.find_or_create_user(user_id, username, default_sort_key)
         return forum_user.user.pk
 
     @staticmethod
@@ -1563,14 +1352,6 @@ class MySQLBackend(AbstractBackend):
         raise ValueError("Comment doesn't have the thread.")
 
     @staticmethod
-    def get_user(user_id: str) -> dict[str, Any] | None:
-        """Return user from user_id."""
-        try:
-            return ForumUser.objects.get(user__pk=int(user_id)).to_dict()
-        except ObjectDoesNotExist:
-            return None
-
-    @staticmethod
     def get_thread(thread_id: str) -> dict[str, Any] | None:
         """Return thread from thread_id."""
         try:
@@ -1750,59 +1531,13 @@ class MySQLBackend(AbstractBackend):
 
     @staticmethod
     def update_user(user_id: str, data: dict[str, Any]) -> int:
-        """
-        Updates user info and ForumUser fields.
-
-        Args:
-            user_id: ID of the user to update.
-            data: Dictionary containing updated user info and ForumUser fields.
-        """
-        try:
-            user = User.objects.get(id=user_id)
-            forum_user = ForumUser.objects.get(user=user)
-        except ObjectDoesNotExist:
-            return 0
-
-        if "username" in data:
-            user.username = data["username"]
-        if "email" in data:
-            user.email = data["email"]
-        if "default_sort_key" in data:
-            forum_user.default_sort_key = data["default_sort_key"]
-        if "read_states" in data:
-            # Remove all existing ReadState objects for this user
-            ReadState.objects.filter(user=user).delete()
-            # Insert new ReadState objects from data['read_states']
-            for state in data["read_states"]:
-                last_read_times = state.get("last_read_times", {})
-
-                for thread_id, dt in last_read_times.items():
-                    thread = CommentThread.objects.get(pk=thread_id)
-                    read_state, _ = ReadState.objects.get_or_create(
-                        user=user, course_id=thread.course_id
-                    )
-
-                    LastReadTime.objects.update_or_create(
-                        read_state=read_state,
-                        comment_thread=thread,
-                        defaults={
-                            "timestamp": dt,
-                        },
-                    )
-
-        user.save()
-        forum_user.save()
-        return 1
+        """Update user info and ForumUser fields."""
+        return ForumUser.update_user(user_id, data)
 
     @staticmethod
     def replace_username_in_all_content(user_id: str, username: str) -> None:
         """Replace the username of a Django user."""
-        try:
-            user = User.objects.get(pk=user_id)
-            user.username = username
-            user.save()
-        except User.DoesNotExist as exc:
-            raise ValueError("User does not exist") from exc
+        ForumUser.replace_username_in_all_content(user_id, username)
 
     @staticmethod
     def get_thread_id_by_comment_id(parent_comment_id: str) -> str:
@@ -1909,74 +1644,22 @@ class MySQLBackend(AbstractBackend):
 
     @staticmethod
     def get_users(**kwargs: Any) -> list[dict[str, Any]]:
-        """
-        Retrieves a list of users in the database based on provided filters.
-
-        Args:
-            kwargs: The filter arguments.
-
-        Returns:
-            A list of users.
-        """
-        forum_users = ForumUser.objects.filter(**kwargs)
-        sort_key = kwargs.get("sort_key")
-        if sort_key:
-            forum_users = forum_users.order_by(sort_key)
-
-        result = [user.to_dict() for user in forum_users]
-        return result
+        """Retrieves a list of users in the database based on provided filters."""
+        return ForumUser.get_users(**kwargs)
 
     @staticmethod
     def get_user_sort_criterion(sort_by: str) -> dict[str, Any]:
-        """
-        Get sort criterion based on sort_by parameter.
-
-        Args:
-            sort_by (str): The sort_by parameter.
-
-        Returns:
-            A dictionary representing the sort criterion.
-        """
-        if sort_by == "flagged":
-            return {
-                "course_stats__active_flags": -1,
-                "course_stats__inactive_flags": -1,
-                "username": -1,
-            }
-        elif sort_by == "recency":
-            return {"course_stats__last_activity_at": -1, "username": -1}
-        else:
-            return {
-                "course_stats__threads": -1,
-                "course_stats__responses": -1,
-                "course_stats__replies": -1,
-                "username": -1,
-            }
+        """Get sort criterion based on sort_by parameter."""
+        return ForumUser.get_user_sort_criterion(sort_by)
 
     @classmethod
     def get_paginated_user_stats(
         cls, course_id: str, page: int, per_page: int, sort_criterion: dict[str, Any]
     ) -> dict[str, Any]:
         """Get paginated user stats."""
-        users = User.objects.filter(
-            Q(course_stats__course_id=course_id)
-            & Q(course_stats__course_id__isnull=False)
-        ).order_by(
-            *[f"-{key}" for key, value in sort_criterion.items() if value == -1],
-            *[key for key, value in sort_criterion.items() if value == 1],
+        return ForumUser.get_paginated_user_stats(
+            course_id, page, per_page, sort_criterion
         )
-
-        paginator = Paginator(users, per_page)
-        paginated_users = paginator.page(page)
-
-        forum_users = [
-            ForumUser.objects.get(user_id=user_id)
-            for user_id in paginated_users.object_list
-        ]
-        return {
-            "pagination": [{"total_count": paginator.count}],
-            "data": [user.to_dict(course_id=course_id) for user in forum_users],
-        }
 
     @staticmethod
     def get_contents(**kwargs: Any) -> list[dict[str, Any]]:
